Split the Delta Lake product test suite per Databricks runtime
