+++ .github/workflows/ci.yml
@@ -683,7 +683,9 @@ jobs:
             # suite-4 does not exist
             - suite-5
             - suite-azure
-            - suite-delta-lake-databricks
+            - suite-delta-lake-databricks73
+            - suite-delta-lake-databricks91
+            - suite-delta-lake-databricks104
             - suite-gcs
           exclude:
             - config: default
@@ -708,9 +710,19 @@ jobs:
               ignore exclusion if: >-
                 ${{ secrets.GCP_CREDENTIALS_KEY != '' }}
 
-            - suite: suite-delta-lake-databricks
+            - suite: suite-delta-lake-databricks73
               config: hdp3
-            - suite: suite-delta-lake-databricks
+            - suite: suite-delta-lake-databricks73
+              ignore exclusion if: >-
+                ${{ secrets.DATABRICKS_TOKEN != '' }}
+            - suite: suite-delta-lake-databricks91
+              config: hdp3
+            - suite: suite-delta-lake-databricks91
+              ignore exclusion if: >-
+                ${{ secrets.DATABRICKS_TOKEN != '' }}
+            - suite: suite-delta-lake-databricks104
+              config: hdp3
+            - suite: suite-delta-lake-databricks104
               ignore exclusion if: >-
                 ${{ secrets.DATABRICKS_TOKEN != '' }}
